SmallChat
SmallChat is a peer to peer UDP-based protocol created by Valentino
Giudice initially designed to create a local network chat without
needing to be connected to the Internet. Now, it can actually also be
used for communicating with hosts outside the network. Despite it was
designed for chat, the protocol can also be used in other scenarios.